move subpackages to npm top-level package
+ spotlessApply
