+++ lib-extra/build.gradle
@@ -24,14 +24,6 @@ dependencies {
 // we'll hold the core lib to a high standard
 spotbugs { reportLevel = 'low' } // low|medium|high (low = sensitive to even minor mistakes)
 
-test {
-	useJUnit {
-		excludeCategories 'com.diffplug.spotless.extra.nodebased.NpmTest'
-	}
-}
+test { useJUnit { excludeCategories 'com.diffplug.spotless.extra.nodebased.NpmTest' } }
 
-task npmTest(type: Test) {
-	useJUnit {
-		includeCategories 'com.diffplug.spotless.extra.nodebased.NpmTest'
-	}
-}
+task npmTest(type: Test) { useJUnit { includeCategories 'com.diffplug.spotless.extra.nodebased.NpmTest' } }
