+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/PrettierFormatterStep.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.npm.prettier;
+package com.diffplug.spotless.extra.npm;
 
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
@@ -28,10 +28,6 @@ import com.diffplug.spotless.FormatterFunc;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.Provisioner;
 import com.diffplug.spotless.ThrowingEx;
-import com.diffplug.spotless.extra.npm.NpmConfig;
-import com.diffplug.spotless.extra.npm.NpmFormatterStepStateBase;
-import com.diffplug.spotless.extra.npm.prettier.options.PrettierOptions;
-import com.diffplug.spotless.extra.npm.wrapper.*;
 
 public class PrettierFormatterStep {
 
@@ -55,7 +51,7 @@ public class PrettierFormatterStep {
 			super(stepName,
 					provisioner,
 					new NpmConfig(
-							readFileFromClasspath(PrettierFormatterStep.class, "package.json"),
+							readFileFromClasspath(PrettierFormatterStep.class, "prettier-package.json"),
 							"prettier"),
 					buildDir,
 					npm);
