+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/ReflectiveObjectWrapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.npm.wrapper;
+package com.diffplug.spotless.extra.npm;
 
 import static java.util.Objects.requireNonNull;
