+++ lib-extra/src/main/java/com/diffplug/spotless/extra/npm/TsFmtOptions.java
@@ -13,13 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.extra.npm.tsfmt;
+package com.diffplug.spotless.extra.npm;
 
 import java.io.Serializable;
 import java.util.Optional;
-
-import com.diffplug.spotless.extra.npm.wrapper.NodeJSWrapper;
-import com.diffplug.spotless.extra.npm.wrapper.V8ObjectWrapper;
 
 public class TsFmtOptions implements Serializable {
